feat: Throw together the basics to re-write the application
This aims to simplify the legacy application that was written over a
decade ago. We're keeping things basic here. As we only need this to
manage our own CV.

Instead of storing the CV entries in a data store, we've gone for using
YAML and JSON. This allows us to still focus on writing our CV and not
have to think about the layout of it.

Signed-off-by: Yomi Colledge <baphled@boodah.net>